refactor [#108] 이름 변경: `ImageInput` -> `AvatarInput`
