+++ src/components/input/index.ts
@@ -1,4 +1,4 @@
-export { default as ImageInput } from "./ImageInput";
+export { default as AvatarInput } from "./AvatarInput";
 export { default as Input } from "./input";
 export { default as InputBox } from "./InputBox";
 export { default as TextField } from "./TextField";
